requirements.txt
hypothesis
numpy
pandas
scipy
pytest-benchmark
